Added TimeManager and Context
TimeMamanger is both made in Eventloop and Window, and stored inside context.
Context is passed in the MainGame.run(), and is supposed to be passed around.
Curernt debug log is from the TestScene.
There is a sleep(100) in both window loop AND eventloop.
(this is for checking deltatime against a set number)